requirements.txt
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
pycryptodome==3.15.0
requests==2.28.1
soupsieve==2.3.2.post1
tuya-connector-python==0.1.2
urllib3==1.26.13
websocket-client==1.4.2
